backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-speech
httpx
sqlalchemy 
pymysql 
boto3
PyJWT
python-decouple
python-dotenv
requests
pymysql
pandas
cryptography
torch
finance-datareader
torchinfo
plotly
scikit-learn
matplotlib
gunicorn
node
ragas
sentence-transformers
pydantic
PyMuPDF
faiss-cpu
transformers
tokenizers
langchain-openai
langchain 
openai
pydantic<1.10.10
haystack-ai
ragas-haystack
langchain-community
openpyxl